Creados repositories, inicializacion base de datos y corregida la DDD

+++ polaflix/src/main/java/es/unican/mmt883/polaflix/repository/FacturaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FacturaRepository extends JpaRepository<Factura, Integer> {
+public interface FacturaRepository extends JpaRepository<Factura, Long> {
 }
