+++ polaflix/src/main/java/es/unican/mmt883/polaflix/repository/SerieRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SerieRepository extends JpaRepository<Serie, Integer> {
+public interface SerieRepository extends JpaRepository<Serie, Long> {
 }
